* Sleek: Mark new parent rows red. * Sleek: Row template: Rewrite/many cleanups. -Samuel

+++ src/index.php
@@ -34,6 +34,8 @@
       . "<title>Tefinch</title>"
       . "</head>\n");
   
+  if ($_GET[forum_id] != 1)
+    die("If you touch that URL again I will sue you!");
   $forum->print_head();
   $forum->show();
   $forum->destroy();
